Some cleanups and compatibility fixes
- Fixed some compatibility issues with other mods
  We now use `OutputPerStack` as a scaling factor
  We no longer hard-code `MaxFuel` for generators
- Cleaned up and refactored update routine a bit
- Improved and fixed `EnumGamePrefs` patching

+++ Harmony/ocbElectricityUtils.cs
@@ -36,15 +36,15 @@ namespace OCB
         public static int chargePerBattery = ChargePerBatteryDefault;
 
         // Get solar cell power by quality
-        static public ushort GetCellPowerByQuality(int quality)
+        static public float GetCellPowerByQuality(int quality)
         {
-            return (ushort)((double)powerPerPanel * (double)Mathf.Lerp(0.5f, 1f, (float)quality / 6f));
+            return powerPerPanel * Mathf.Lerp(0.5f, 1f, quality / 6f);
         }
 
         // Get discharge power by battery quality
-        static public ushort GetDischargeByQuality(int quality)
+        static public float GetDischargeByQuality(int quality)
         {
-            return (ushort)((double)powerPerBattery * (double)Mathf.Lerp(0.5f, 1f, (float)quality / 6f));
+            return powerPerBattery * Mathf.Lerp(0.5f, 1f, quality / 6f);
         }
 
         // Get charging power by battery quality
